refactor(telegram): move intermediate/outbound send helpers to intermediates.rs, verbatim (#471 phase 1)
Fourth mechanical cut: flush_intermediates, the last-intermediate
footer append, the rich-first intermediate send, the rate-limit
retrying wrapper and send_html_or_plain move to intermediates.rs.
Pure move, pub(crate) bumps, glob re-export keeps all paths stable.
